git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/mips-3_4-rewrite-branch@58434 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/gcov-io.h
@@ -231,7 +231,7 @@ struct function_info
 /* Information about a single object file.  */
 struct gcov_info
 {
-  unsigned version;	        /* expected version number */
+  unsigned long version;        /* expected version number */
   struct gcov_info *next;	/* link to next, used by libgcc */
 
   const char *filename;		/* output file name */
